Different levels of solidness for buildings
This feature makes it possible to distinguish between solid buildings a warhead cannot shoot through and solid buildings a warhead should try to shoot through, first destroying the building and then the real target behind it. This is comparable to walled in buildings, were units also attack the building no matter that a wall is in the way.

Only enemy buildings' solidity will be considered. Units will not try to fire through allied buildings.

This can be used to differentiate between solid buildings that are in the way but usually have to be destroyed, like gates, and solid buildings like bunkers that really are impenetrable and where it would be smarter to reposition the unit. Also, it allows for example units with elite weapons to go on a rampage, ignoring everything around them.

For historical reasons, SolidLevel is in artmd.ini, like SolidHeight. This might change in the future.

[BuildingType]SolidLevel= (integer, defaults to 1)
The solidity of this building.

[Warhead]SolidLevel= (integer, defaults to 0)
The highest level of solidity this warhead will consider as to be easily destructible when targeting. If a solid building with a solidity level of this or lower is in the way to the actual target, the unit will fire anyhow, first destroying all obstacles in its way. If is is lower than the building's level, then the unit cannot fire and has to reposition itself to fire.

The default is the following: All buildings are more solid than warheads are able to shoot through them. Solid buildings count as unconditionally solid, just like in previous versions of Ares.