Moved SolidHeight and SolidLevel to rulesmd
For backwards compatibility, SolidLevel is still read from SolidLevel, but the one in rulesmd takes precedence. This only applies if they aren't used at the same time, though. As the value is read from artmd for every new INI read, it would override the setting from rulesmd when reading the game mode, and the game mode setting when reading a scenario or map.

SolidLevel didn't exist yet, so moving it comes at no cost.