Fully recoded AudioIndex_LoadSample
Now loading bagged and loose files is unified. Also reproduced a piece of code that supports external audio folders. This is not used by the game, and for now the block is deactivated using an #ifdef.

Note that the commented out block does not play well with loose files, and overwriting ExternalFile will leak the current file.