Damage Air Threshold
[Warhead]DamageAirThreshold= (integer - leptons, defaults to 0)
If the warhead detonates more than this many leptons above the ground, units in the air will be affected. Otherwise, units in the air are not damaged. Use -1 to even damage air units when the warhead detonates on the ground, or a high positive value like 9999 to disable damaging air units.