Consider bridges when finding out whether a warhead detonated on the ground
[General]DamageAirConsiderBridges= (boolean, defaults to no)
Whether a warhead detonating above a bridge will consider the bridge's height for finding out whether units in air are to be damaged. If no, a warhead detonating on a bridge will be considered to be four levels (416 leptons) in the air, if yes, it will be 0 leptons above the ground and by default not trigger damaging units in the air.