Customizable Movies List
The new system does not rely on two counters, but unlocks each movie individually. The unlocked state is saved in ra2md.ini's new [UnlockedMovies] section.

In moviesmd.ini, add all movie filenames to the [Movies] list. Movies will appear on the in-game list in the same order as in this list. Only if the list is empty does the game default to the original behavior using the hardcoded movie lists and the two counters.

[MovieFilename]Description= (CSF label)
The name this movie will appear under in the list.

[MovieFilename]Unlockable= (boolean, defaults to yes)
Whether this movie has to be unlocked first before appearing in the list. If no, the movie will always appear in the list, like the intro movie.