Fixed a minor flaw in the index checks
The checks only worked half the time, when Index was too small, that is the GenericPrerequisite index was too large and thus out of bounds. If Index was too large, and thus the GenericPrerequisite index became negative, the check erroneously succeeded.

Also made the Index converted to a GenericPrerequisite index a dedicated variable.