Support arbitrary types to trigger prerequisite checks
Whenever a units change owner, are killed, put or discovered, the construction options will be rechecked. Because Alternative Prerequisites use a different system than buildings, removing alone does not invalidate build options, as units in transporters also count as present. Neither does attacking a unit with a temporal weapon invalidate anything, as units can't be factories (which would affect build options if temporally attacked).

For each unit type, the check whether it is a Generic Prerequisite is only done once, and the result is cached. Not sure whether this is so important, because the checks don't happen too frequently, but iterating over all alternate list items for each generic prerequisite seemed like a stupid thing to do.