Template Update 49: Changed how Promotable<T> is parsed.
If the base tag is defined, all three promotion level settings are set to that value, and afterward the three specific tags are read for each individual level.

The name of the single-flag can be provided separately. If null, the base name is reused with "%s" being replaced by an empty string. If the last character then becomes a dot, it is removed in that case.