Make the reload times for each dock on an UnitReload=yes building independent
This fixes the issue that reload times seem to vary for all planes and the plane that landed first is substantially faster (it requires one more interval of ReloadRate).

This happened if the first docker landed a bit earlier than the others. The building goes into repair mission, which puts the docker to sleep, if it isn't doing so already. If it does, one ammo is restored. If the other dockers land only after the first docker, they are not put to sleep and will thus be put to sleep after the first ReloadRate interval finishes. At this point, the first docker will get its first round of ammo restored already.

This does not happen any more. The mission will update each frame, not just every ReloadRate * 900 frames, thus units go to sleep immediately, and will reload frame-percise individually.