Related to #894922: Aircraft can dock again
Aircraft can dock to any structure that is in Dock, even if they don't have UnitRepair=yes set. This requirement was kinda weird anyhow, because that doesn't work together with UnitReload=yes, which is set on the Airforce Command HQ.

A minor change from the original code: UnitRepair was used to switch over to enter and no-enter cursors, and only buildings that were UnitRepair=yes got one of those cursors. Now, The structure has to be in Dock, otherwise the select cursor is used.