RadarJamming optimization
The update mechanism was inefficient. It continuously unjammed even buildings that are unjammable, which is not a no-op. It needlessly registered each building's owner for rechecking their radars every time any Radar Jammer on the map updated.

Now, buildings changing owners are manually unjammed, and the next call to Update() will recreate the jamming if the building is still eligible. Likewise, Radar Jammers changing owners unregister from all buildings.