JammerClass's Jam() and Unjam() overhauled
Both methods are now doing more work, as both check whether the jammer is or isn't in the map already. Both insert and erase iterate over all items anyhow, thus complexity does not increase. In the case of no change, there's not even runtime overhead. In the case of a change, there is, though.

RecheckRadar is now only set if there is a change, and only if that change is relevant. This should, how shall we say, even ze score, ja?