Removed some of Splits and Airburst code by solving the parts containing literals. Now it's all gone.
The removed code makes sense, though:
dir is the direction on a flat map the bullet is fired in. radians is used to get the x and y components from the unit circle, thus, sin_rad and cos_rad are factors and their squares sum up to 1. Likewise, threeHalfPi is the firing angle (not tested: horizontal 0, up Pi/2, backwards Pi, down 3Pi/2), sin_factor and cos_factor used as factors for the parallel and orthogonal components.