Cache relevant data for SW Targeting.
This iterates over the buildings and technos only once, not once per techno considered a potential target. The Designator and Inhibitor calculations have been changed to work on the squared distances, which gets rid of a few conversions.

This way is much more efficient, though the complexity is the same. This is because it doesn't recalculate the same data all the time. In case of manual targeting this is a bit slower because it doesn't exit early, but it should be a huge win on the auto-targeting side.