Saboteurs inspired by Dune 2000, without timed cloak
[InfantryType]Saboteur= (boolean, defaults to no)
Whether this infantry can blow up one sabotagable building. The building will be destroyed as if C4 was placed, and the Saboteur is consumed in this process. Requires Infiltrate=yes. Not supported together with vehicle hijacking, CanDrive=yes, C4 and Occupier=yes logics.

A building counts as sabotagable if it is owned by an enemy player and has CanC4=yes set.

Sabotage does give the "enter" cursor, not the "sabotage" one. This cannot be changed easily without changing the implementation. The reason is that the unit is consumed by the process, like a spy, thief, or engineer.

One hook has been removed, as it is included in the saboteur logic.