Use three global vectors in GetTypeToProduce, so the memory is reused.
All three vectors are shared between all instantiations of GetTypeToProduce, and thus their capacity will be the maximum of all there type list counts. As the original vectors were locals, their data is not shared between function calls and this is safe to do, as long as the vectors are reset on function re-entry.

Now the 100-unit fix is almost as efficient (amortized) as the fixed arrays in the vanilla game.