ReloadAmount and EmptyReloadAmount
[TechnoType]ReloadAmount= (integer, defaults to 1)
The number of rounds restored after each reload interval. Negative values are supported, as long as EmptyReloadAmount is equal to Ammo.

[TechnoType]EmptyReloadAmount= (integer, defaults to ReloadAmount)
The number of rounds restored after reload interval when there is no ammo left. If 0 or negative, the unit will not reload. This does not require EmptyReload to be set.