Weapons can spend more or less than 1 round of Ammo
[Weapon]Ammo= (integer, defaults to 1)
The rounds of ammunition needed to fire this weapon. If fewer than the required rounds are present, the unit will not fire and wait for the ammo to reload.

This feature does not support aircraft. Units that do ManualReload=yes will still target enemies and wait for reloading, which never happens. The assumption |ammo != 0 -> can fire| has not been updated in all places yet.