Relay Log and and DebugLog calls to the LogFlushed methods
This gets rid of the original Log function pointer. This gets rid of an indirect call to a game function through that pointer. This gets rid of a hook being executed each time something is logged. The hook would call LogWithVArgs anyhow, which is now called directly.

The story is shorter with DevLog. It is almost the same as LogFlushed, but doesn't register a parsing error.