Another Container update: use a non-virtual base class for the pointer to pointer map.
ContainerMap wraps the ContainerMapBase and provides type safety. unique_ptr usage has been removed again, in favor of a single pair of new and delete, contained in the Container methods. Nothing changes to the outside world.

One internal change is that clearing the Container will now leak memory. That's not really a feature, though. Clearing should only happen if the Container is empty and thus all extensions have been destroyed already when their owners were destroyed. If a leak warning is logged, we have another problem, and the leak itself is just a symptom.

On the pro side: this should get rid of a lot of needlessly duplicated code, because there were over a dozen unordered_maps that essentially all did the same thing, and each had their own yet common memory management functions. Now they are instantiated only once, in the new _Container.cpp file.

This might make retrieving extensions faster, because there are not a dozen different functions to walk binary trees, with only the types being different. There's now only one find method, which hopefully stays in the instruction cache a bit longer, instead of being removed to make place for an identical function when the next extension object of another Container is queried.