Let the AI build more than one building of each type
Additional buildings are not build one after another as a block right after the original building, but instead mix randomly with the structures planned to be built later in the build list.

[BuildingType]AIBuildCounts= (list of integers, defaults to 1,1,1)
The number of buildings the AI will build of this type depending on the difficulty. Order is hard to easy. Each value must be equal to or greater than 1.

[BuildingType]AIExtraCounts= (list of integers, defaults to 0,0,0)
The maximum number of extra buildings the AI can build of this type depending on the difficulty. Order is hard to easy. A random number of buildings will be build in addition to the ones from the normal build count. Each value must be equal to or greater than 0.