Creating an own iterator does not optimize out well in case the method called in a loop is empty.
Even if there is absolutely nothing to do in the body of a range for loop (because the function call is inlined and the function is empty), code is still generated to get the first and the last, then increment first until it reaches end.

The "solution" is ugly. It uses the standard iterators, but casts them as if they were iterators of a differently typed map. They return the correct values with the correct type, but a real adapter would be a bit more standard compliant.

Quick and dirty test results: Inheriting from the iterator would work, but apparently as soon as any member of non-trivially-constructible type is added, it's back at the unwanted behavior again.