Fixed the shp compression 1 issue with too big shapes
The issue is a too small static buffer which is used as a threshold for pixel alpha values or something like that. The buffer is used for each line, so it creates vertical lines.

A vector is used that dynamically grows. This is also needed, because reading beyond the static array's bounds is possible without crash, though it's technically undefined. This is not guaranteed to work with an object on the heap.