Drop out early if the announced pointer is not interesting at all. This speeds up Ares pointer invalidation a lot.
It's still rather clumsy, but easier to maintain than than to register each ExtData class for certain abstract types. Everything has been set up as it was previously. Some invalidations might not be necessary, and if they are removed, some of this could be deleted again.

Watch out: if you need to invalidate something for sure, make sure its type does return false in InvalidateExtDataIgnorable().