Sort the damage groups using insertion sort, and only sort as many as actually needed.
There is no change if MaxAffect is larger than the number of damage groups. In the common case of MaxAffect=1, the algorithm is reduced in complexity to a simple std::min_element.

Also, the algorithm is stable while the last one was not, which shouldn't have mattered for the same binaries and with the same data which does not take the addresses of objects into consideration.