Finer grained ProtectedDriver or increased KillDriver resistance
[TechnoType]ProtectedDriver.MinHealth= (double - percentage, defaults to 0.0 if ProtectedDriver=yes, to 1.0 otherwise)
The minimum health below which the driver of this unit can be killed. If the health is above this, the driver cannot be killed. If KillDriver.KillBelowPercent= is also defined on the warhead, the minimum of the two values is used, that is, this tag can make a unit more resistant against driver killing weapons.