Refinement of CanPassiveAquire
The following tags refine the way CanPassiveAquire works when a human-controlled unit is in guard mode. It does not affect AI-controlled units. Units can be made to effectively hold fire when in guard mode (and optionally cloakable), yet pick their own targets when in area guard mode. This was the default in Tiberian Sun, and it made Stealth Tanks not give away their position when an enemy unit came into range.

Note that the second tag is more specific than the first as it means cloakable AND in guard mode. It is not independent of the first. CanPassiveAquire.Guard is the tag that makes units always hold their fire in guard mode, CanPassiveAquire.Cloak can be used for units that should hold fire only if they gained the veteran or elite ability.

[TechnoType]CanPassiveAquire.Guard= (boolean, defaults to yes)
Whether this unit can acquire targets on their own when in guard mode. If set to no, the unit will hold fire, but can be made to acquire own targets by putting it in area guard mode. Only affects human controlled units, and does not work on buildings. Requires CanPassiveAquire=yes.

[TechnoType]CanPassiveAquire.Cloak= (boolean, defaults to yes)
Whether this unit can acquire targets on their own when the unit is cloakable and in guard mode. If set to no, the unit will hold fire if cloakable and in guard mode, but can be made to acquire own targets by putting it in area guard mode. Only affects human controlled units, and does not work on buildings. Requires CanPassiveAquire=yes. CanPassiveAquire.Guard=no overrides this tag.

Units count as cloakable if they have Cloakable=yes set or were awarded cloak ability as veteran or elite ability. The unit does not have to be actually cloaked; also, passive mechanisms like Cloak Generators will not be considered.

Note the spelling of both tags. Ares keeps the spelling error of CanPassiveAquire to stay consistent.