Related to #1489302: Loaded trucks won't give crates
A register is set in the skipped code which hasn't been recreated. It was used as fifth parameter to Pathfinding_Find() to get the best crate position, and thus the function returned an empty CellStruct. This is the invalid cell, and the code caught it and didn't place the crate.