Initial Payload (Generals-style Troop Crawlers)
[TechnoType]InitialPayload.Types= (list of TechnoTypes)
The types of the initial payload an object of this type is created with. Types can be mentioned more than once. If there are fewer items in InitialPayload.Nums= list, the last Num is used to define the count. If there are no numbers, the count is 1.
BuildingTypes and AircraftTypes are not valid. Buildings are only allowed to have InfantryType payload. Units having initial payload are not supported, because this could create infinite loops.

[TechnoType]InitialPayload.Nums= (list of integers)
The number of times each type from InitialPayload.Types= will be added. If this list contains fewer items than the types, the last number is used as count. If this list is empty, then each count is 1.