Custom Missiles support weapons instead of warhead / damage settings
[AircraftType]Missile.Weapon= (WeaponType, defaults to none)
The weapon used to deliver damage. If set, the weapon is used to control projectile, warhead, damage and bright settings for a short-lived bullet that detonates immediately. If not set, Missile.Damage and Missile.Warhead are used to create the explosion. Requires Missile.Custom=yes.

[AircraftType]Missile.EliteWeapon= (WeaponType, defaults to none)
Same as Missile.EliteWeapon, but overrides Missile.EliteDamage and Missile.EliteWarhead for missiles fired by elite units. Requires Missile.Custom=yes.

Note that missiles will not fall back to Missile.Weapon in case Missile.EliteWeapon is not set. Missile.EliteDamage and Missile.EliteWarhead are used.

Primary and ElitePrimary are not used to not also have a death weapon by default, which can lead to double explosions when the missile detonates a warhead that destroys the missile before it is removed from the game.