Specific Passengers
[TechnoType]Passengers.Allowed= (list of TechnoTypes)
The list of allowed passenger types. If this contains at least one type, all other types are not allowed to enter this object. Only supported on VehicleTypes and AircraftTypes. If you want to prevent all units from entering, set this to either a dummy unit or a BuildingType.

[TechnoType]Passengers.Disallowed= (list of TechnoTypes)
The list of types that are not allowed to enter this object. Only supported on VehicleTypes and AircraftTypes.