Template Update 53: Added default copy and move constructors and assignment operators to each Valueable<T>-derived class.
Certain user defined special functions disable implicit generation of the other special functions. This varies from C++ standard to standard. Now the classes are explicit about that they define.

Copy and move constructors allow Nullable<T> and derived to also copy the empty state. Previously, the Nullable  copied from could be converted to its value, which was then used to create a new Nullable. Empty values are still implicitly convertible, and constructing a Nullable that way always set the copy to non-empty.

Not that this happened, but the problem with implicitly generated functions is that you don't see when they change.