Template Update 54: Added assignment operator template to each Valueable<T>-derived class.
This gets rid of some casts, and disallowed assignments that should be possible. Also, this allows to specify all one-parameter constructors as explicit. Previously, assignment was done by implicitly converting to Valueable<T>, then copying.

Changed the type of the Valueable<T> assignment operator to not use && on a template type, because that is a catch-all, which also matches non-const lvalue references and thus prevents the default assignment operators to work.

Because now every class implements its own assignment operator, don't call Set but set all values directly. A consequence is that it's no longer possible to call set in a polymorphic way.