Template Update 60: Devirtualized all ValueableVector<T>-derived templates
Load() and Save() are never called through pointer or reference to base, so the virtual specifier could just be removed. The Read() method was the only thing that called the virtual Split() function, and by rewriting Read() for all four types [Valueable|Nullable][|Idx]Vector, it could be removed along with the non-virtual parse method. With all virtual methods gone, the destructor could be removed.