Plane count will not reset if Count tag not present in map INI
When the rules defined several planes and the game mode INI or map file had a section for the ParaDrop SW but didn't restate the Counts, the default plane was truncated to 1, then this cascaded to sides and houses.

Now, the default count is only used if the value is not already in the map, otherwise the size of the vector in the map is used.