Template Update 62: Devirtualized Valueable<T>-derived templates
This should have been done some time ago already. There is no need to have a virtual destructor, because there are no virtual methods left. This really drives home a point: there is no longer a virtual function table, which means that the alignment will not be forced to be at least 4. This means that for example Valueable<bool> is now only one byte where it used 8 bytes previously.

(Template Update 56 was used twice, so 61 is skipped to keep this useless count up tp date.)