Disallow entering a unit
[TechnoType]NoManualEnter= (boolean, defaults to no)
Whether units will not get an enter or no-enter cursor when on this object. If yes, the select cursor is used, as if this unit is not a transport.

This does not prevent units from entering by other means, like scripts, though. Also, the AI will not respect this setting. Use the specific passenger feature to prevent this.