Revert "Initial commit of Jekyll setup."
This reverts commit f811c528226f008bc8688f5549f697ff7514183d.

Looking at the work required to create a Jekyll page with a custom theme compared to what we want to achieve, especially considering that the page hasn't sustantially changed in the past ten years, I've decided that using GitHub's Jekyll magic isn't worth the effort.
The existing page is almost static anyway.
I'll convert that and if any volunteer wants to do new, amazing things with the site later, that's what it's on GitHub for.